requirements.txt
discord.py==2.3.2 # Or the latest stable version you prefer
python-dotenv==1.0.1 # Or latest
Pillow==11.2.1 # Or latest
SQLModel==0.0.18 # Or latest
PyYAML==6.0.1 # Or latest (for your .yml files)
aiosqlite==0.20.0